Fogs/requirements.txt
django
psycopg2
django-debug-toolbar
dadata
django_extensions
djangorestframework